Removed next-themes; refactored dashboard components; added dark mode support
