+++ layouts/dashboard/mainSiteDashboard.tsx
@@ -1,7 +1,8 @@
 
 import * as Prisma from '@prisma/client';
 import { ClockIcon, CollectionIcon, HomeIcon } from '@heroicons/react/solid';
-import DashboardLayout, { SidebarLink } from '.';
+import { SidebarLink } from 'components/dashboard/sidebar';
+import DashboardLayout from '.';
 
 const sidebar: SidebarLink[] = [
   { name: 'Home', href: '/', icon: HomeIcon, route: '/_sites/[site]' },
@@ -14,7 +15,7 @@ interface MainSiteDashboardLayoutProps {
   children?: React.ReactNode;
 }
 const MainSiteDashboardLayout: React.FC<MainSiteDashboardLayoutProps> = ({ community, children }) => {
-  return <DashboardLayout community={community} sidebar={sidebar}>{children}</DashboardLayout>;
+  return <DashboardLayout community={community} sidebarMenuItems={sidebar}>{children}</DashboardLayout>;
 };
 
 export default MainSiteDashboardLayout;
