+++ package.json
@@ -41,7 +41,6 @@
     "micro": "^9.3.4",
     "micro-cors": "^0.1.1",
     "next": "^12.1.6",
-    "next-themes": "^0.2.0",
     "nexus": "^1.3.0",
     "nexus-prisma": "0.34.0",
     "postcss-import": "^14.0.2",
