+++ pages/_app.tsx
@@ -1,7 +1,6 @@
 import 'styles/tailwind.css';
 import { ApolloProvider } from '@apollo/client';
 import { Provider } from 'react-redux';
-import { ThemeProvider } from 'next-themes';
 import { UserContextWrapper } from 'context/UserContext';
 import { store } from 'app/store';
 import client from 'context/ApolloContext';
@@ -10,15 +9,13 @@ import type { AppProps } from 'next/app';
 function MyApp({ Component, pageProps }: AppProps) {
 
   return (
-    <ThemeProvider>
-      <ApolloProvider client={client}> {/* Apollo first so react context can use graphql */}
-        <Provider store={store}>
-          <UserContextWrapper>
-            <Component {...pageProps} />
-          </UserContextWrapper>
-        </Provider>
-      </ApolloProvider>
-    </ThemeProvider>
+    <ApolloProvider client={client}> {/* Apollo first so react context can use graphql */}
+      <Provider store={store}>
+        <UserContextWrapper>
+          <Component {...pageProps} />
+        </UserContextWrapper>
+      </Provider>
+    </ApolloProvider>
   );
 }
